Revert "fix(route): 更新地址Update xygg.js (#10059)"
This reverts commit 77d7439a3b391ab938a66551a570f761791fc29e.
