+++ lib/router.js
@@ -3037,7 +3037,7 @@ router.get('/hfut/tzgg', lazyloadRouteHandler('./routes/universities/hfut/tzgg')
 // router.get('/darwinawards/all', lazyloadRouteHandler('./routes/darwinawards/articles'));
 
 // 四川职业技术学院
-// router.get('/scvtc/xygg', lazyloadRouteHandler('./routes/universities/scvtc/xygg'));
+router.get('/scvtc/xygg', lazyloadRouteHandler('./routes/universities/scvtc/xygg'));
 
 // 华南理工大学土木与交通学院
 router.get('/scut/scet/notice', lazyloadRouteHandler('./routes/universities/scut/scet/notice'));
