--- lib/v2/scvtc/maintainer.js
@@ -1,3 +1,0 @@
-module.exports = {
-    '/xygg': ['nczitzk'],
-};
