--- lib/v2/scvtc/radar.js
@@ -1,13 +1,0 @@
-module.exports = {
-    'scvtc.edu.cn': {
-        _name: '四川职业技术学院',
-        '.': [
-            {
-                title: '学院公告',
-                docs: 'https://docs.rsshub.app/university.html#si-chuan-zhi-ye-ji-shu-xue-yuan',
-                source: ['/ggfw1/xygg.htm', '/'],
-                target: '/scvtc/xygg',
-            },
-        ],
-    },
-};
