--- lib/v2/scvtc/router.js
@@ -1,3 +1,0 @@
-module.exports = (router) => {
-    router.get('/xygg', require('./xygg'));
-};
